Merge BCrypt main haaran kanssa. 
- Salasanan enkryptaus.
- Käyttäjät näkevät vain omat tietonsa.
- Kategorioiden luominen

+++ src/main/java/fi/opiskelijan/budjettilaskuri/web/MenoController.java
@@ -7,10 +7,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import java.security.Principal;
 
+import fi.opiskelijan.budjettilaskuri.domain.Kayttaja;
 import fi.opiskelijan.budjettilaskuri.domain.Meno;
 import fi.opiskelijan.budjettilaskuri.domain.Kategoriakuukausi;
 import fi.opiskelijan.budjettilaskuri.repository.KategoriaRepository;
+import fi.opiskelijan.budjettilaskuri.repository.KayttajaRepository;
 import fi.opiskelijan.budjettilaskuri.repository.MenoRepository;
 
 @Controller
@@ -22,23 +25,38 @@ public class MenoController {
     @Autowired
     private KategoriaRepository kategoriaRepository;
 
+    @Autowired
+    private KayttajaRepository kayttajaRepository;
+
     @GetMapping("/menot")
-    public String listaMenot(Model model) {
-        model.addAttribute("menot", menoRepository.findAll());
+    public String listaMenot(Model model, Principal principal) {
+        String username = principal.getName();
+
+        model.addAttribute("menot", menoRepository.findByKayttajaUsername(username));
+        
+        model.addAttribute("kategoriat", kategoriaRepository.findByKayttajaUsername(username));
+        
         model.addAttribute("uusiMeno", new Meno());
-        model.addAttribute("kategoriat", kategoriaRepository.findAll());
+        
         return "lisaa-meno";
     }
 
     @PostMapping("/menot")
-    public String lisaaMeno(@ModelAttribute Meno meno) {
+    public String lisaaMeno(@ModelAttribute Meno meno, Principal principal) {
+        Kayttaja kayttaja = kayttajaRepository.findByUsername(principal.getName());
+
+        meno.setKayttaja(kayttaja);
+
         menoRepository.save(meno);
         return "redirect:/menot";
     }
 
     @PostMapping("/menot/{id}/poista")
-    public String poistaMeno(@PathVariable Long id) {
-        menoRepository.deleteById(id);
+    public String poistaMeno(@PathVariable Long id, Principal principal) {
+        Meno meno = menoRepository.findById(id).orElse(null);
+        if (meno != null && meno.getKayttaja().getUsername().equals(principal.getName())) {
+                    menoRepository.deleteById(id);
+                }
         return "redirect:/menot";
     }
